Create fruits classification
The project's primary goal is to build, train, and evaluate a CNN model that can accurately classify images of fruits into their respective categories. This type of project is commonly used as an introductory example in computer vision and deep learning to showcase the process of building and deploying image classification models.